+++ Dyalect/Parser/Model/NodeType.cs
@@ -27,6 +27,7 @@
         Continue,
         Return,
         Yield,
+        YieldBreak,
         Throw,
 
         Directive,