+++ Docs/Changes.md
@@ -4,6 +4,8 @@
   * Add: _Add an ability to create types with constructors in foreign modules_ ([Issue #395](https://github.com/vorov2/dyalect/issues/395)).
   * Add: _Callable objects_ ([Issue #389](https://github.com/vorov2/dyalect/issues/389), [docs](https://github.com/vorov2/dyalect/wiki/Functions#callable)).
   * Add: _String should support slicing operator_ ([Issue #396](https://github.com/vorov2/dyalect/issues/396)).
+  * Add: _Add a support for `yield break`_ ([Issue #399](https://github.com/vorov2/dyalect/issues/399), [docs](https://github.com/vorov2/dyalect/wiki/Iterators#return)).
+  * Add: _`return` shouldn't be allowed in iterators_ ([Issue #400](https://github.com/vorov2/dyalect/issues/400)).
 
 # 0.17.5
   * Change: _Code optimization and refactoring._